Add files via upload
A Jupyter notebook with the completed assignment from a Coursera course. It creates a fake album cover by grabbing a random image from a website as the album image. The fake album title and band title are drawn on the album cover. The titles are grabbed from titles of randomly selected Wikipedia pages, so some web scraping is involved.